+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Meio que só fiz uma changelog no update final heh.
  
 ### Fixed
 
-## [2.0] - 2024-12-13
+## [2.0.0] - 2024-12-13
  
 ### Added
 - MAIS 5 PERGUNTAS NOVAS!
@@ -32,3 +32,20 @@ Meio que só fiz uma changelog no update final heh.
  
 ### Fixed
 - Problemas com input.
+
+## [2.1.0] - 2024-12-26
+ 
+### Added
+- Sistema de Pontuação!
+- Pular perguntas!
+- Opção de ajuda, para saber como funciona o jogo!
+- DIFICULDADES NORMAL E DÍFICIL!
+- Na dificuldade normal, se você errar perderá pontos, e você terá direito a pular uma pergunta.
+- Na dificuldade difícil, você não terá direito a pular nenhuma pergunta. E se você perder, reiniciará o jogo.
+
+### Changed
+- Adicionado comentários de como funciona o código.
+- Timer de 2 segundos entre textos.
+ 
+### Fixed
+-